Move some of the sequitur related structures
